+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Settings/EpisodeSettings.h
@@ -18,4 +18,6 @@ struct CARLA_API FEpisodeSettings
 
   UPROPERTY(EditAnywhere, BlueprintReadWrite)
   bool bNoRenderingMode = false;
+
+  TOptional<double> FixedDeltaSeconds;
 };